Move geometry template tests into separate files
